Added push notification to the UWP project (that logic was missing since it was not part of the original template).
Also added more description in error code to better explain common failure when debugging in adroid emulator.  If gsf package was not added to emulator then you will get this error but not it will be called out in error making it easier to understand exactly what is going on.